improving compilation and installation process
